+++ lib/logic/cubit/home_screen_state.dart
@@ -1,13 +1,16 @@
 part of 'home_screen_cubit.dart';
 
 class HomeScreenState extends Equatable{
-
+  final bool animated;
+  const HomeScreenState({required this.animated});
   @override
-  List<Object> get props => [];
+  List<Object> get props => [animated];
 
 }
 
 class HomeScreenInitial extends HomeScreenState {
+  const HomeScreenInitial({required super.animated});
+
 
   @override
   List<Object> get props => [];
@@ -17,7 +20,7 @@ class HomeScreenInitial extends HomeScreenState {
 class HomeScreenIncremented extends HomeScreenState {
   final int value;
 
-  HomeScreenIncremented({required this.value});
+  const HomeScreenIncremented({required super.animated, required this.value});
 
   @override
   List<Object> get props => [value];